+++ src/provider/nominatim/NominatimReverseQuery.ts
@@ -1,18 +1,35 @@
 import { ReverseQuery, ReverseQueryObject } from "query";
 
+type Shape = "geojson" | "kml" | "svg" | "text";
+
 export interface NominatimReverseQueryObject extends ReverseQueryObject {
   readonly zoom?: number;
+  readonly shape?: Shape;
+  readonly shapeThreshold?: number;
 }
 
 export default class NominatimReverseQuery extends ReverseQuery {
   private readonly zoom?: number;
 
+  private readonly shape?: Shape;
+
+  private readonly shapeThreshold?: number;
+
   protected constructor({
     zoom,
+    shape,
+    shapeThreshold,
     ...reverseQueryObject
   }: NominatimReverseQueryObject) {
     super(reverseQueryObject);
     this.zoom = zoom;
+    if (shape && !["geojson", "kml", "svg", "text"].includes(shape)) {
+      throw new Error(
+        'The "shape" parameter can only have the following values: "geojson", "kml", "svg", "text".'
+      );
+    }
+    this.shape = shape;
+    this.shapeThreshold = shapeThreshold;
   }
 
   public static create(
@@ -25,6 +42,8 @@ export default class NominatimReverseQuery extends ReverseQuery {
     return {
       ...super.toObject(),
       zoom: this.zoom,
+      shape: this.shape,
+      shapeThreshold: this.shapeThreshold,
     };
   }
 
@@ -34,5 +53,21 @@ export default class NominatimReverseQuery extends ReverseQuery {
 
   public getZoom(): undefined | number {
     return this.zoom;
+  }
+
+  public withShape(shape: Shape): NominatimReverseQuery {
+    return new NominatimReverseQuery({ ...this.toObject(), shape });
+  }
+
+  public getShape(): undefined | Shape {
+    return this.shape;
+  }
+
+  public withShapeThreshold(shapeThreshold: number): NominatimReverseQuery {
+    return new NominatimReverseQuery({ ...this.toObject(), shapeThreshold });
+  }
+
+  public getShapeThreshold(): undefined | number {
+    return this.shapeThreshold;
   }
 }